Lib will raise a SyntaxError if invalid file
Not perfect, but should catch a lot of use cases. Really,
whoever uses this code should check syntax ahead of time
so we can just pull from the stream instead of having to read
it all in to syntax check it, defeats the purpose of allowing
streams in the first place.