No longer print exception message on stderr
An exception message used to print to stderr, but I realized that
makes it more difficult to drop into an editor because it means
the program will print and stderr will print, looking like it
worked successfully. So you would have to send stderr to /dev/null,
but that meant that when it actually blew up, you wouldn't see the
error message and be able to report the issue. So, exception messages
aren't printed on stderr anymore.