Add cuke to expose Parser error
Not implemented because not released by Parser yet.
We could point the Gemfile at it, but because Parser
does not include their generated Ruby files in the
git repo, we then have to cd to that dir and build
the project in order to get Bundler to use it correctly.

Fuck that, amirite?