"refactor" CommentLines
Disliked how it was passing all the values back in the array, and the
naming was making me uncomfortable. So pulled it out into more methods.
Which I actually like worse, but it is useful for the other classes.

IDK, maybe the other possibility is that there is a common other object,
which just finds commentable lines, and then that is used by
CommentLines, and Align* we'll see.