Bump version to 2.1.3
If you were on 2.1.x, the raeason minor changed rather than patch
is because the way the_matrix.rb works changed fairly significantly.
So if you overrode that file, you'll need to look at it and figure out
what to do for your thing (see #24 for an in depth explanation).

Also, added a dependency on parser. This should allow it to work on
rubinius (not tested on CI yet because some cukes are failing for
formatting reasons).