Consumer assumes garbled events are just stdout
If an event is produced that the consumer can't understand,
it makes the assumption that it is output. This is because
the process could have been `exec`'d or something. And
rather than just blowing up, would be really nice to actually
record and show the output from the new process