Code always ends in a newline
Normalize the code to always end in a newline internally.

This is because having to juggle the maybe/maybenot-ness of the newline
all over the place winds up infecting code all over the place with
distracting complexity that is easy to get wrong and is often unclear what
the right thing to do is.

Maintain behaviour by checking to see whether the original source had a newline
before printing it and then print the final result without it if that is the case.