Rename Binary::InterpretFlags -> Binary::Engine
Was trying to figure out what to do about the feature envy. I wanted to
push it down into InterpretFlags, but then its doing a lot more than
just interpreting the flags. I tried pulling it all up into the binary,
but it got cluttery up there, and I disliked that all that functionality
was then coupled to the binary's progression and side effects.
So I realized it definitely belongs below Binary. Not completely sure it
all belongs here (rather than in, say a separate object), but it
definitely goes better here than anywhere else at present, so I'm going
to put that behaviour into this class. And thus, it needs a better name.
Engine seemed adequate, it's the thing that does the work and provides
the context for the binary to know what to do and when, without getting
all caught up in the details.