Wow. Fixed BEGIN/END.
This is really an artifact of the fact that with the event stream,
we no longer need to wrap the entire file. This allows us to simply not wrap
the expression itself (we can wrap its contents, still), and it will
naturally wind up working.

I'm kinda disappointed that I don't have an issue to go close :/