Rename event observers back to handlers, but put under a namespace
Handler is a better name afterall, b/c I'm going ot just have the toplevel
return the handler and not the return status. Was just doing that to try and
maintain the toplevel interface, but this is v3, I've got tons of breaking changes.
The model is fundamentally different now (event stream vs go do a thing and return
me a data structure of the events. So callers will just have to get the result
off the handler if they use the default. 1 extra method call, not that big of a deal,
and works much better with the internal structure. (ie this whole return_value
thing on the handlers is just dumb.