Uhm... sigh, legit spent 3+ hours trying to get this to work.
SimpleCov has like 50 places you can enter it, all with some implicit
knowledge about the environment. It keeps overriding exit statuses,
but not in reliably reproducable ways (e.g. I run the suite, it blows up,
I run the one spec, it's fine). It calls Kernel.exit, overriding
the error that SiB needs to record. I tried to clean it up, but moving
one line to another in a way that should absolutely not matter, winds
up causing its tests to fail. This is because it stores all its sate
in singletons, so somehow something happening in the interim was going
and mutating it.

IDK, coverage API looks pretty simple, and SimpleCov breaks down as soon
as you deviate from its expectation, so if I really care about this, I'll
probably just write my own and see if I can't use their formatter for
inspecting it.

I will give it this, though: when I didn't shoot for 100% and thus
didn't try to get it to consolidate its result across processes, then it
didn't blow up, and the info I got from that was useful.

But srsly, kill off the singletons.