Add --ignore-unknown-flags to allow compatibility with multiple versions of SiB
* Add flag --ignore-unknown-flags
* Mention it in the help screen.
* Some rewording between "flag" and "options"... in general,
  my language seems to be confused about when to use which.

Code which integrates with SiB (e.g. editor integrations) have difficulty
supporting changing interfaces. This flag allows the integration to specify
that unknown options not be treated as errors (e.g. typos). This will allow
a situation such as:

```
SiB v100 has flags
  √ --flag1
  √ --flag2
  X --flag3

Sib v200 has flags
  X --flag1
  √ --flag2
  √ --flag3
```

Currently, the integration would have to check which version it is using
to determine which flags to set. Otherwise, this would happen:

```

$ seeing_is_believing _100_ --flag1 --flag2 --flag3
Error: --flag3 is not a flag, see the help screen (-h) for a list of options

$ seeing_is_believing _200_ --flag1 --flag2 --flag3
Error: --flag1 is not a flag, see the help screen (-h) for a list of options
```

Now, though, it can use the flag to tell each version to ignore the flags the don't know,
and thus call SiB with a set of flags that will be compatible across multiple versions:

```
$ seeing_is_believing _100_ --flag1 --flag2 --flag3 && echo $?
0

$ seeing_is_believing _200_ --flag1 --flag2 --flag3 && echo $?
0
```

Some flags got renamed, but I have to use the deprecated versions
(see features/deprecated-flags.feature) because they work on both
v2 and v3. Would be better to be able to specify to ignore the new flag,
and then set them both.