The TCPServer uses port 0, which will dynamically allocate an available port
This is part of #94

Remove the --port flag. Because we choose an available port, the user shouldn't need
to specify this. Additionally, we don't want the user to depend on it because this
part of the code has been very volatile with major implementation changes like fifteen
times. By removing it, we prevent users from depending on it, which is preferrable due
to its volatility.