Ensure descendants get killed instead of orphaned, remove ichannel, improve tests
It should hopefully no longer be possible for tests to pass when they leave orphan processes.
Also many more of the tests will now work on JRuby and Windows, the only tests that use fork
now are the tests that test fork. Since we aren't using fork, ichannel isn't helpful anymore,
so remove it from dev deps

Part of #94
Addresses a number of concerns from #92