Missing exit statuses are converted to 1 (closes #100)
I'm not totally sure this is the right thing for it to do, but a segfault is the only way
I know of to invoke this situation, and a segfault is printable. So until I get some info
that proves this is the wrong thing to do, we're just going to give it a normal exit status.

It's the easiest thing to do, and it's more correct in this one case.  ¯\_(ツ)_/¯