CI Configuration Tweaks
Rakefile: Since both CI servers invoke `rake install` on the install
step (see .travis.yml and appveyor.yml) the CI task doesn't need to
depend on that one.

Travis (Unix): Run tests against Ruby 2.4 also. This requires us to turn
off a cuke on JSON behaviour related to #46 The JSON specification now
allows for a JSON document to be a string. IOW, quirks mode is now the
default mode. Thus, this test does not illustrate the old error. See the
diff for links to explanations.

AppVeyor (Windows): Windows branch is merged now, so we can run against
the master branch. I don't completely get the settings here, I stole them
from RSpec. I'm guessing its saying that there are two branches that
github creates for a PR, one of them is named "something-maintenance",
which we can match with a regex, so we match that one and master causing
master to be tested and the maintenance branch to be tested and skipping
the other incidental branch.

AppVeyor: Also remove the comment about the "bundle exec" being weird
since it's the same settings I'm using for Travis (presumably it was
intentional at the time, at the very least it doesn't seem to be hurting
anything)