Add --local-cwd flag, which CDs to the file's dir before running
Not fond of the implementation in EvaluateByMovingFiles, but there's a lot of stuff
going on around me and so it's a bit difficult to concentrate. I think those if
statements could be moved up to #initialize, though, and then it could always set
the child process's cwd (rather than changing the current process's cwd) and it
could always use a relative filename. Basically, it could probably get a lot more
elegant, but I'm having difficulty concentrating and need to go do shit.