Remove 2.4 and 2.5 from Appveyor
They deadlock on one of the tests, but other than that are okay.
I don't have time right now to deal w/ that, but I want to push
this version, b/c it's super annoying to use SiB right now.
I'll have to add them back later and either not run those tests
on Windows on 2.4/2.5, or I'll have to figure out why that happens
and fix it.