Fix #136
The spy was accidentally returning itself from `inspect`,
which then called `to_str`, returning itself again. It
would then marshal dump the Spy object, and send it to the
consumer. The consumer exploded when unmarshaling b/c it
couldn't the class only existed on the producer side.

This uses `String.===` to make sure that the result of
`.to_str` is actually a string, and falls back to other
inspect strategies if not.

We also add `String.===` to the refinements since SiB
depends on it.