Switch from childprocess gem to Kernel#spawn
SiB has been broken on the M1 (Apple Silicon) for a long time,
unnecessarily. I reported that it would work if they just updated
the gem to understand that that architecture exists:
https://github.com/enkessler/childprocess/issues/176

But they haven't gotten back to it. So I figured I'd try out
`Kernel#spawn` again. Turns out, it works, on OS X, anyway.
We'll see what happens on CI.